// src/app/types/PlanetsData.ts
export interface ISearchData {
  count: number;
  next: string | null;
  previous: string | null;
  results: IPlanetData[];
}

export interface IPlanetData {
  climate: string;
  created: string;
  diameter: string;
  edited: string;
  films: string[];
  gravity: string;
  name: string;
  orbital_period: string;
  population: string;
  residents: string[];
  rotation_period: string;
  surface_water: string;
  terrain: string;
  url: string;
}